+++ app/src/main/java/com/isodev/personalaibuddyui/ui/theme/Color.kt
@@ -9,3 +9,9 @@ val Pink80 = Color(0xFFEFB8C8)
 val Purple40 = Color(0xFF6650a4)
 val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)
+
+// Color
+val ColorBlack = Color(0xFF000000)
+val ColorBlack2 = Color(0xFF121111)
+val ColorWhite = Color(0xFFFFFFFF)
+val ColorYellow = Color(0xFFC6F432)
