+++ app/src/main/res/values/colors.xml
@@ -5,10 +5,4 @@
     <color name="purple_700">#FF3700B3</color>
     <color name="teal_200">#FF03DAC5</color>
     <color name="teal_700">#FF018786</color>
-    <color name="black">#FF000000</color>
-    <color name="white">#FFFFFFFF</color>
-    <color name="yellow">#FFC6F432</color>
-    <color name="black2">#121111</color>
-
-
 </resources>
